pyproject.toml
[project]
name = "outlook-ai-assistant"
version = "0.1.0"
description = "AI-powered email management agent for Outlook using Claude"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "anthropic>=0.78.0",           # Claude API client (tool use, structured output)
    "msal>=1.34.0",                # Microsoft authentication (device code flow)
    "requests>=2.32.0",            # HTTP client for Graph API (sync operations)
    "aiohttp>=3.9.0",              # Async HTTP client for Graph API
    "fastapi>=0.128.0",            # Web framework for review UI
    "uvicorn>=0.40.0",             # ASGI server
    "jinja2>=3.1.5",               # HTML templates
    "pyyaml>=6.0.2",               # Config file parsing
    "pydantic>=2.10.0",            # Config schema validation
    "pydantic-settings>=2.0.0",    # Settings management
    "apscheduler>=3.11.0,<4.0.0",  # Job scheduling (3.x stable API)
    "click>=8.3.0",                # CLI framework
    "rich>=14.0.0",                # Console output formatting (progress bars, tables)
    "structlog>=25.1.0",           # Structured JSON logging
    "python-dateutil>=2.9.0",      # Date parsing and manipulation
    "aiosqlite>=0.22.0",           # Async SQLite
    "regex>=2024.0.0",             # Regex with timeout support (CRITICAL for security)
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",               # For FastAPI test client
    "ruff>=0.9.0",                 # Linting and formatting
    "pytest-cov>=4.0.0",           # Coverage reporting
]

[project.scripts]
assistant = "assistant.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/assistant"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"src/assistant/web/routes.py" = ["B008"]  # Depends() in defaults is standard FastAPI
"src/assistant/web/dependencies.py" = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v"
